Reorganized the package structure by potential module breakdown
