+++ engine/src/test/java/org/destinationsol/systems/DamageSystemTests/NonNegativeHealthTest.java
@@ -15,9 +15,9 @@
  */
 package org.destinationsol.systems.DamageSystemTests;
 
-import org.destinationsol.components.Health;
+import org.destinationsol.health.components.Health;
 import org.destinationsol.entitysystem.EntitySystemManager;
-import org.destinationsol.events.DamageEvent;
+import org.destinationsol.health.events.DamageEvent;
 import org.destinationsol.modules.ModuleManager;
 import org.junit.Before;
 import org.junit.Test;
